requirements.txt
asgiref==3.4.1
backcall==0.2.0
colorama==0.4.4
decorator==5.1.0
Django==3.2.7
flake8==3.9.2
importlib-metadata==4.8.1
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.10.0
python-ldap==3.3.1
pytz==2021.1
sqlparse==0.4.2
traitlets==5.1.0
typing-extensions==3.10.0.2
wcwidth==0.2.5
zipp==3.5.0
